fixed problem with trustAll setting in config.
fixes #3
trustAll was being checked for null but it would never be
null unless set to null in config. In our case trustAll
would eval into an instance of a ConfigObject even if it
was not set. Changed it to check for it to be that or null
then set trustAll = true if it was running in non prod